requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
insightface>=0.7.0
openpyxl>=3.1.0
onnxruntime>=1.15.0
tk>=0.1.0
pillow>=9.0.0